Hub/requirements.txt
-i https://pypi.org/simple
astroid==2.2.5
autopep8==1.4.4
entrypoints==0.3
flake8==3.7.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
numpy==1.17.2
paho-mqtt==1.4.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
six==1.12.0
typed-ast==1.4.0 ; implementation_name == 'cpython'
wrapt==1.11.2
